requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==5.0
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==4.3.1
django-filter==23.5
django-jazzmin==2.6.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
Markdown==3.5.1
pycparser==2.21
python-decouple==3.8
pytz==2023.3.post1
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.4
tzdata==2023.3
